food_stories/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.1
django-4-jet==1.0.9
django-modeltranslation==0.18.7
idna==3.4
oauthlib==3.2.2
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
typing_extensions==4.4.0
urllib3==1.26.12
